Moves some theme-related actions to their own proper Action classes
